+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 * Support `ISWAP`, `SISWAP`, and their adjoints.
   [(#114)](https://github.com/PennyLaneAI/pennylane-cirq/pull/114)
 
+* Support a variety of operations raised to a power.
+  [(#115)](https://github.com/PennyLaneAI/pennylane-cirq/pull/115)
+
 ### Breaking changes
 
 ### Improvements
@@ -12,6 +15,10 @@
 * Pass all the qubits as `qubit_order` parameter to force 
   the simulator not to ignore idle qubits. 
   [(#111)](https://github.com/PennyLaneAI/pennylane-cirq/pull/111) 
+
+* Removed the unnecessary `CPhase` and `ISWAP` operations from the
+  plugin's custom support because PennyLane supports them.
+  [(#115)](https://github.com/PennyLaneAI/pennylane-cirq/pull/115)
 
 ### Documentation